Add play styles to the agent using sepperate classes. Playstyles are defined within the CalculatorHanlder and every playstyle implements Calculator. This might change to extending some day. Now two play styles are added random and cheater.
Random uses a random number generator to determine which stack and how many should be removed.
Cheater changes the stacks to 1-0-0 in one turn.

Fifteenstack will print in error output when 0 is taken from the stack.
Same goes for adding (- a number using the take function)

Agent will congratulate when lose. Also will check if the opposite player sends an gameover stack and call him a cheater.